requirements.txt
Flask==2.3.3
numpy==1.25.2
rasterio==1.3.2
Pillow==9.5.0
gunicorn==20.1.0